Revert "Move FileExtensions into PluginSpecification"
This reverts commit e87fd0c35d5e5d7b26ac37dff189a8e8fef844ec.
